Made a bunch of changes (#21)
I rewrote a bunch of stuff to use classes (though it can probably be better). I made rooms polygons instead of squares. I find the closest polygon for the room instead of just finding a polygon that the device is in. I add history for measurements and locations. The measurements used to calculate the position is a weighted average of the measurements based on time.  I changed the error function to be inversely proportional to the distance squared.